pyproject.toml
[project]
name = "blacksheep-proxy"
version = "0.1.0"
description = "API Proxy with BlackSheep and JWT authentication"
authors = [{name="Luis Forni"}]
requires-python = ">=3.11"

dependencies = [
    "blacksheep>=2.0.5",
    "uvicorn[standard]>=0.30.6",
    "pydantic>=2.7.4",
    "pydantic-settings>=2.4.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
